requirements.txt
annotated-types==0.5.0
anyio==3.7.1
apifairy==1.3.0
apispec==6.0.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.7.0
cffi==1.15.1
click==8.1.6
comm==0.1.2
cryptography==39.0.0
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
Flask==2.2.2
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==5.1.5
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.2.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.14.0
PyJWT==2.6.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
tomli==2.0.1
tornado==6.2
traitlets==5.8.1
typing_extensions==4.7.1
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webargs==8.2.0
websockets==11.0.3
Werkzeug==2.2.2
